LPS-61461 Move StrutsPortlet related code to StrutsPortletAuthTokenWhitelist
